move Info util, clarify popover positioning docs
