+++ src/components/forms/checkbox.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Preview, Props, Source } from "@storybook/addon-docs/blocks";
 import { withKnobs, text, boolean } from '@storybook/addon-knobs';
-import Info from '../../util/Info.jsx';
+import { Info } from '../util/storybook';
 
 import Checkbox from './Checkbox';
