+++ src/components/popovers/popover.stories.mdx
@@ -7,6 +7,7 @@ import {
   boolean,
 } from '@storybook/addon-knobs';
 import { arrayToOptions } from '../util/storybook';
+import { Info } from '../util/storybook';
 
 import Popover, {
   VALID_ALIGNMENTS,
@@ -59,6 +60,11 @@ Base popover component. Bring your own trigger and popover content.
 - Arrows. They aren't supported, but we can add them in the future.
 
 <Props of={Popover} />
+
+<Info type="info">
+  When setting `position` and `alignment` props, the positioning engine will prefer the props you
+  provide, but may move the popover content to avoid edges of the viewport. This behavior is intentional.
+</Info>
 
 <Preview>
   <Story name="Styled popover content" height={100}>
